// src/Moryx.Operators.Web/src/app/api/strict-http-response.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpResponse } from '@angular/common/http';

/**
 * Constrains the http response to not have the body defined as `T | null`, but `T` only.
 */
export type StrictHttpResponse<T> = HttpResponse<T> & {
  readonly body: T;
}